USB Serial fixes
