+++ hal/src/stm32f2xx/usb_settings.h
@@ -19,9 +19,9 @@
 #ifndef USB_SETTINGS_H_
 #define USB_SETTINGS_H_
 
-#define USB_TX_BUFFER_SIZE              128  /* Total size of IN buffer:
+#define USB_TX_BUFFER_SIZE              129  /* Total size of IN buffer:
                                                 APP_RX_DATA_SIZE*8/MAX_BAUDARATE*1000 should be > CDC_IN_FRAME_INTERVAL */
-#define USB_RX_BUFFER_SIZE              256
+#define USB_RX_BUFFER_SIZE              257
 
 #define USB_SERIAL_USERSPACE_BUFFERS    1
